BC-10090 hp5-proxy: return 404 on root path
To prevent users from accessing h5p.org directly through the proxy,
this change introduces a new location block in the NGINX configuration
that returns a 404 Not Found error for requests to the root path (`/`).

This ensures that only specific, intended paths are proxied, while
the root path is blocked, addressing the unintended access issue.